+++ scanomatic/image_analysis/image_basics.py
@@ -81,7 +81,7 @@ def Quick_Scale_To_im(path=None, im=None, source_dpi=600, target_dpi=150,
 
         try:
 
-            im = load_image_to_numpy(path, dtype=np.uint8)
+            im = load_image_to_numpy(path)
 
         except:
